pyproject.toml
[tool.poetry]
name = "blog-html-to-md"
version = "0.1.0"
description = ""
authors = ["Phil Whelan"]

[tool.poetry.dependencies]
python = "^3.8"
lxml = "^4.5.2"
python-dateutil = "^2.8.1"
html2markdown = "^0.1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
